Added input for transactions

+++ BudgetApp/Transaction.swift
@@ -3,16 +3,16 @@ import UIKit
 
 class Transaction {
     var title: String
-    var amount: Int
+    var amount: Double
     var description: String
     
-    init(title: String, amount: Int) {
+    init(title: String, amount: Double) {
         self.title = title
         self.amount = amount
         self.description = "No description."
     }
     
-    init(title: String, amount: Int, description: String) {
+    init(title: String, amount: Double, description: String) {
         self.title = title
         self.amount = amount
         self.description = description
